back/requirements.txt
asgiref==3.8.1
boto3==1.35.49
botocore==1.35.49
dj-database-url==2.3.0
Django==4.2.16
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.2.3